move common fluid crafting code into CraftingFluidBlock
